gvp programming completed

+++ plug-ins/hard/rpspmc_hwi_structs.h
@@ -293,13 +293,13 @@ typedef struct{
 	gint32    i,j;           /**<10,11: loop counter(s) =RO/Zero */
 	gint32    ptr_next;      /**<12: next vector (relative to VPC) until --rep_index > 0 and ptr_next != 0 =WR */
         gint32    ptr_final;     /**<13: next vector (relative to VPC), =1 for next Vector in VP, if 0, probe is done =WR */
-	gint32    f_du;          /**<14: U (bias) stepwidth (32bit) =WR */
-	gint32    f_dx;          /**<16: X stepwidth (32bit) =WR */
-	gint32    f_dy;          /**<18: Y stepwidth (32bit) =WR */
-	gint32    f_dz;          /**<20: Z stepwidth (32bit) =WR */
-	gint32    f_dx0;         /**<22: X0 (offset) stepwidth (32bit) =WR */
-	gint32    f_dy0;         /**<24: Y0 (offset) stepwidth (32bit) =WR */
-	gint32    f_dz0;         /**<26: Z0 (offset)  =WR */
+	double    f_du;          /**<14: U (bias) stepwidth in Volts =WR */
+	double    f_dx;          /**<16: X stepwidth in Volts =WR */
+	double    f_dy;          /**<18: Y stepwidth in Volts =WR */
+	double    f_dz;          /**<20: Z stepwidth in Volts =WR */
+	double    f_dx0;         /**<22: X0 (offset) stepwidth in Volts =WR */
+	double    f_dy0;         /**<24: Y0 (offset) stepwidth in Volts =WR */
+	double    f_dz0;         /**<26: Z0 (offset) stepwidth in Volts =WR */
 } PROBE_VECTOR_GENERIC;
 
 /**
